feat(inventory,sales): unidad de medida y cantidad decimal
00036 — quita el límite que impedía vender por peso o medida.

`quantity` era `int` en las CUATRO tablas del flujo de mercancía
(inventory_item, entry_line, exit_line, sale_line), así que 12,5 g no se
podía ni representar. Salió diseñando la fundición, pero no es una función
de oro: NINGUNA compraventa podía vender nada por peso ni por medida — oro
por gramo, cable por metro, lo que fuera.

Curiosamente el lado de empeño ya lo hacía bien desde 00005
(`contract_item.weight_grams` es numeric). Era el inventario el corto.

DOS PIEZAS
· `product.unit` — enum (`unit`, `gram`, `kilogram`, `meter`, `liter`), no
  texto libre: con texto libre terminan conviviendo "gr", "grs", "gramo" y
  "gramos" y no hay forma de sumar ni mostrar nada consistente. El backend
  expone también la abreviatura para que front, comprobantes y reportes
  digan todos lo mismo.
· `quantity` pasa a numeric(14,3). Al miligramo, y en la misma familia que
  el dinero para no meter floats en ningún lado.

TRES REGLAS QUE ESTO HABILITA
· Un producto medido en `unit` RECHAZA fracciones. Media cadena no existe:
  una cantidad fraccionaria ahí es un error de digitación —una coma donde
  iba un punto— y registrarlo deja stock imposible que nadie nota hasta que
  el conteo físico no cuadre. En una venta cuesta más caro todavía: cobra un
  total que no corresponde a nada.
· Reponer conserva la unidad del producto EXISTENTE. Si una compra nueva
  pudiera imponer la suya, bastaría dejar el selector en su valor por
  defecto para reinterpretar en silencio el stock anterior.
· La unidad no se cambia con lotes registrados: doce unidades no son doce
  gramos, y cambiarla reinterpretaría stock, ventas y valorización sin que
  nada lo advierta. Mismo criterio que impide cambiar el TIPO de una cuenta.

EL DINERO. `subtotal` ahora pasa por `quantize`: con cantidad decimal, el
producto puede arrastrar milésimas y el redondeo a dos decimales tiene que
ocurrir ANTES de sumar, o el total del recibo no cuadra con la suma de sus
líneas.

Y un bug que destapó la migración: `AppError` llevaba un `Decimal` crudo en
`details` al no haber stock para un egreso, así que ese 400 reventaba al
serializar y salía como 500 sin mensaje.

Se puede aplicar antes del deploy: ensanchar int a numeric es lossless y el
código viejo tolera la ventana (Pydantic convierte Decimal integral a int, y
todo el dato existente es entero).

Tests: 272 (seis nuevos).

+++ app/modules/sales/repository.py
@@ -80,7 +80,7 @@ async def insert_sale_line(
     company_id: UUID,
     sale_id: UUID,
     item_id: UUID,
-    quantity: int,
+    quantity: Decimal,
     unit_price: Decimal,
     unit_cost: Decimal,
     subtotal: Decimal,
